examples/plugins/example-tool/pyproject.toml
[project]
name = "hopeit_agents-example-tool"
version = "0.1.0"

description = "Example MCP-compatible tool implemented as a hopeit.engine plugin"
dynamic = ["readme"]

dependencies = [
    "hopeit.engine>=0.27.0",
    "hopeit-agents.mcp-server",
]

license = { text = "Apache 2" }
authors = [
    { name = "hopeit_agents contributors", email = "opensource@hopeit.com.ar" }
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: AsyncIO",
]

[project.urls]
"Homepage" = "https://github.com/hopeit-git/hopeit.agents"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
readme = { file = ["README.md"], content-type = "text/markdown" }
